backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langgraph
openai
faiss-cpu
beautifulsoup4
networkx
pypdf
arxiv
python-dotenv
tiktoken
